Rename validation schemas folder for clarity
