--- schemas/index.ts
@@ -1,13 +1,0 @@
-import * as z from "zod";
-
-export const LoginSchema = z.object({
-  username: z.string(),
-  password: z.string()
-});
-
-export const RegisterSchema = z.object({
-  email: z.string().email(),
-  username: z.string(),
-  password: z.string(),
-  confirmPassword: z.string()
-});
